requirements.txt
# PolyPrinting2026 - Polymarket Trading Bot Dependencies
# Python 3.10+ required

# =====================================================
# Polymarket Official Client
# https://github.com/Polymarket/py-clob-client
# =====================================================
py-clob-client>=0.17.0

# =====================================================
# Cryptocurrency Exchange Integration
# https://github.com/ccxt/ccxt
# =====================================================
ccxt>=4.2.0

# =====================================================
# Web & API
# =====================================================
requests>=2.31.0
websocket-client>=1.7.0
aiohttp>=3.9.0

# =====================================================
# Data Analysis
# =====================================================
pandas>=2.1.0
numpy>=1.26.0

# =====================================================
# Configuration & Environment
# =====================================================
pyyaml>=6.0.1
python-dotenv>=1.0.0

# =====================================================
# Ethereum / Web3 (required by py-clob-client)
# =====================================================
web3>=6.15.0
eth-account>=0.11.0

# =====================================================
# Monitoring & Metrics (Optional)
# =====================================================
prometheus-client>=0.19.0

# =====================================================
# Scheduling & Async
# =====================================================
schedule>=1.2.0
asyncio-throttle>=1.0.2

# =====================================================
# Logging & Utilities
# =====================================================
structlog>=24.1.0
colorama>=0.4.6

# =====================================================
# Testing
# =====================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# =====================================================
# Development (optional)
# =====================================================
# black>=24.1.0
# flake8>=7.0.0
# mypy>=1.8.0
